more robust handling of upsertmediaprogress
